Some fixes and cleanup for ExampleDrill, delete MoltenDrill from Old

+++ ExampleMod/Content/Projectiles/ExampleDrillProjectile.cs
@@ -1,6 +1,5 @@
 ﻿using ExampleMod.Content.Dusts;
 using Microsoft.Xna.Framework;
-using System;
 using Terraria;
 using Terraria.Audio;
 using Terraria.ID;
@@ -10,9 +9,14 @@ namespace ExampleMod.Content.Projectiles
 {
 	public class ExampleDrillProjectile : ModProjectile
 	{
+		public override void SetStaticDefaults() {
+			// Prevents jitter when steping up and down blocks and half blocks
+			ProjectileID.Sets.HeldProjDoesNotUsePlayerGfxOffY[Type] = true;
+		}
+
 		public override void SetDefaults() {
-			Projectile.width = 18;
-			Projectile.height = 18;
+			Projectile.width = 22;
+			Projectile.height = 22;
 			Projectile.friendly = true;
 			Projectile.tileCollide = false;
 			Projectile.penetrate = -1;
@@ -25,10 +29,11 @@ namespace ExampleMod.Content.Projectiles
 		// This code is adapted and simplified from aiStyle 20 to use a different dust and more noises. If you want to use aiStyle 20, you do not need to do any of this.
 		// It should be noted that this projectile has no effect on mining and is mostly visual.
 		public override void AI() {
-
 			Player player = Main.player[Projectile.owner];
 
 			Projectile.timeLeft = 60;
+
+			// Animation code could go here if the projectile was animated. 
 
 			// Plays a sound every 20 ticks. In aiStyle 20, soundDelay is set to 30 ticks.
 			if (Projectile.soundDelay <= 0) {
@@ -36,38 +41,42 @@ namespace ExampleMod.Content.Projectiles
 				Projectile.soundDelay = 20;
 			}
 
-			Vector2 myPosition = player.RotatedRelativePoint(player.MountedCenter);
-			if (player.channel) {
-				float speed = player.HeldItem.shootSpeed * Projectile.scale;
-				Vector2 newVelocity = Main.MouseWorld - myPosition;
-				float distanceLength = newVelocity.Length();
-				distanceLength = speed / distanceLength;
-				newVelocity *= distanceLength;
-				if (newVelocity.X != Projectile.velocity.X || newVelocity.Y != Projectile.velocity.Y) {
-					Projectile.netUpdate = true;
-				}
+			Vector2 playerCenter = player.RotatedRelativePoint(player.MountedCenter);
+			if (Main.myPlayer == Projectile.owner) {
+				// This code must only be ran on the client of the projectile owner
+				if (player.channel) {
+					float holdoutDistance = player.HeldItem.shootSpeed * Projectile.scale;
+					// Calculate a normalized vector from player to mouse and multiply by holdoutDistance to determine resulting holdoutOffset
+					Vector2 holdoutOffset = holdoutDistance * Vector2.Normalize(Main.MouseWorld - playerCenter);
+					if (holdoutOffset.X != Projectile.velocity.X || holdoutOffset.Y != Projectile.velocity.Y) {
+						// This will sync the projectile, most importantly, the velocity.
+						Projectile.netUpdate = true;
+					}
 
-				Projectile.velocity = newVelocity;
+					// Projectile.velocity acts as a holdoutOffset for held projectiles.
+					Projectile.velocity = holdoutOffset;
+				}
+				else {
+					Projectile.Kill();
+				}
 			}
-			else
-				Projectile.Kill();
 
 			if (Projectile.velocity.X > 0f) {
 				player.ChangeDir(1);
-			}			
+			}
 			else if (Projectile.velocity.X < 0f) {
 				player.ChangeDir(-1);
 			}
-				
 
 			Projectile.spriteDirection = Projectile.direction;
 			player.ChangeDir(Projectile.direction); // Change the player's direction based on the projectile's own
 			player.heldProj = Projectile.whoAmI; // We tell the player that the drill is the held projectile, so it will draw in their hand
 			player.SetDummyItemTime(2); // Make sure the player's item time does not change while the projectile is out
-			Projectile.Center = myPosition;
-			Projectile.rotation = (Projectile.velocity).ToRotation() + MathHelper.PiOver2;
+			Projectile.Center = playerCenter; // Centers the projectile on the player. Projectile.velocity will be added to this in later Terraria code causing the projectile to be held away from the player at a set distance.
+			Projectile.rotation = Projectile.velocity.ToRotation() + MathHelper.PiOver2;
 			player.itemRotation = (Projectile.velocity * Projectile.direction).ToRotation();
 
+			// Gives the drill a slight jiggle
 			Projectile.velocity.X *= 1f + Main.rand.Next(-3, 4) * 0.01f;
 
 			// Spawning dust
